Move "main" icon set to `src`-level and eliminate duplicative test `icons`
The pattern library had accrued a set of icons it was using and registering.
In addition, there was a separate `test` icon set, and the icons
registered in the pattern library's icon set were useful for use in
tests for components outside of the pattern library.

This commit move the pattern library's `icons` module up to the main
`src` level where it can be used more broadly, and registers these
icons during test bootstrap.

This was needed, e.g., for the forthcoming `Panel` component, which
assumes the availability of the `cancel` icon during tests.

We may elect to change the overall icon setup as we go here, and this
isn't an assertion that this package will be responsible for managing
shared icons, necessarily.
