--- src/pattern-library/icons.js
@@ -1,13 +1,0 @@
-// @ts-nocheck
-
-export default {
-  'arrow-left': require('../../images/icons/arrow-left.svg'),
-  'arrow-right': require('../../images/icons/arrow-right.svg'),
-  cancel: require('../../images/icons/cancel.svg'),
-  check: require('../../images/icons/check.svg'),
-  collapsed: require('../../images/icons/collapsed.svg'),
-  edit: require('../../images/icons/edit.svg'),
-  logo: require('../../images/icons/logo.svg'),
-  profile: require('../../images/icons/profile.svg'),
-  trash: require('../../images/icons/trash.svg'),
-};
