--- test/icons.js
@@ -1,7 +1,0 @@
-/**
- * Set of icons used by the tests
- */
-export default {
-  'arrow-left': require('../images/icons/arrow-left.svg'),
-  'arrow-right': require('../images/icons/arrow-right.svg'),
-};
